+++ tests/GraphQl/Resolver/Stage/ReadStageTest.php
@@ -66,7 +66,7 @@ class ReadStageTest extends TestCase
      */
     public function testApplyDisabled(array $context, $expectedResult): void
     {
-        $operationName = 'query';
+        $operationName = 'item_query';
         $resourceClass = 'myResource';
         $resourceMetadata = (new ResourceMetadata())->withGraphql([
             $operationName => ['read' => false],
@@ -94,7 +94,7 @@ class ReadStageTest extends TestCase
      */
     public function testApplyItem(?string $identifier, $item, bool $throwNotFound, $expectedResult): void
     {
-        $operationName = 'query';
+        $operationName = 'item_query';
         $resourceClass = 'myResource';
         $info = $this->prophesize(ResolveInfo::class)->reveal();
         $context = [
@@ -184,7 +184,7 @@ class ReadStageTest extends TestCase
      */
     public function testApplyCollection(array $args, ?string $rootClass, ?array $source, array $expectedFilters, iterable $expectedResult): void
     {
-        $operationName = 'query';
+        $operationName = 'collection_query';
         $resourceClass = 'myResource';
         $info = $this->prophesize(ResolveInfo::class)->reveal();
         $fieldName = 'subresource';
@@ -201,9 +201,9 @@ class ReadStageTest extends TestCase
         $normalizationContext = ['normalization' => true];
         $this->serializerContextBuilderProphecy->create($resourceClass, $operationName, $context, true)->shouldBeCalled()->willReturn($normalizationContext);
 
-        $this->subresourceDataProviderProphecy->getSubresource($resourceClass, ['id' => 3], $normalizationContext + ['filters' => $expectedFilters, 'property' => $fieldName, 'identifiers' => [['id', $resourceClass]], 'collection' => true])->willReturn(['subresource']);
+        $this->subresourceDataProviderProphecy->getSubresource($resourceClass, ['id' => 3], $normalizationContext + ['filters' => $expectedFilters, 'property' => $fieldName, 'identifiers' => [['id', $resourceClass]], 'collection' => true], $operationName)->willReturn(['subresource']);
 
-        $this->collectionDataProviderProphecy->getCollection($resourceClass, null, $normalizationContext + ['filters' => $expectedFilters])->willReturn($expectedResult);
+        $this->collectionDataProviderProphecy->getCollection($resourceClass, $operationName, $normalizationContext + ['filters' => $expectedFilters])->willReturn($expectedResult);
 
         $result = ($this->readStage)($resourceClass, $rootClass, $operationName, $context);